+++ src/entities/aliens/BaseAlien.js
@@ -16,11 +16,17 @@ export default class BaseAlien extends Phaser.GameObjects.Container {
     constructor(scene, x, y) {
         super(scene, x, y);
         scene.add.existing(this);
+        this.setDepth(45);  // fly above terrain, props, snail, station, and ground items
         this.facing = 'right';
     }
 
     /** Initialise the sprite after subclass has set this.spriteKey */
     _initSprite() {
+        const shadow = this.scene.add.graphics();
+        shadow.fillStyle(0x000000, 0.35);
+        shadow.fillEllipse(2, 10, 44, 14);
+        this.add(shadow);
+
         this.sprite = this.scene.add.image(0, 0, `${this.spriteKey}-right`);
         this.add(this.sprite);
     }